[ttdebug] Display storage are data; key 'S' shows all of them.
* src/ttdebug.c (Storage): New structure.  We need to handle storage
area data specially since we have to make a distinction between
initialized and uninitialized values.
(handle_WS): New function.  Since it is not necessary for the
bytecode interpreter to trace whether a value is uninitialized or
not, we	have to do it manually right before every debugging step.
(RunIns): After executing a step, storage area data different to
before is displayed.
<'?'>: Help updated.
<'S'>: Show storage area data.
<'c'>, <'s'>, <'n'>: Call `handle_WS'.