[graph/win32] Implement smooth resizing.
The window creation and the message loop has to be put in a separate
thread in order to achieve smooth resizing.  Othrewise, the modal
loop cannot be interrupted to update the canvas.  The main thread
also has a loop to receive the reposted messages, to update the
canvas and send WM_PAINT back (UpdateWindow).  After this update
the win32 resizing feels similar to the X11 resizing.