[meson] Add option to disable librsvg dependency.
As discussed in #14.

Current librsvg is used to render OT-SVG glyphs. However, librsvg is far from
simple to build as it requires tons of transitive dependencies, especially
on Windows.

Therefore, we could introduce a configure switch in the meson script to switch
librsvg on or off. This is trivial because the source code itself already
honors a `HAVE_LIBRSVG` definition.

* meson_options.txt: Add `librsvg` option (feature), defaulting to `enabled`.

* meson.build: Honor the option above.