[meson] Add option to disable librsvg dependency.
As discussed in #14.

Currently, 'librsvg' is used to render OT-SVG glyphs.  However, 'librsvg' is
far from simple to build as it requires tons of transitive dependencies,
especially on Windows.

Therefore we introducee a configuration option to switch 'librsvg' on or
off.  This is trivial because the source code itself already honors a
`HAVE_LIBRSVG` macro.

* meson_options.txt: Add `librsvg` option (feature), defaulting to
`enabled`.

* meson.build: Honor the option above.