[ftinspect] Move out `ftgrid` out from `MainGUI` to a new tabbed pane.
For further addition of other FreeType demo programs, we will use a tabbed
pane to organize all views, and the original `ftgrid` view will be the first
one, named "Singular Grid View".

An `AbstractTab` is added for helping the `MainGUI` to notify all sub-views
when settings change. The `AbstractTab` is a pure-virtual class itself.

Settings dedicate to singular grid view are moved to their tab, and the labels
showing current glyph name/index are moved too. These overlay labels are made
transparent to mouse click so it don't block in the way.

* src/ftinspect/maingui.cpp, src/ftinspect/maingui.hpp: Move out all `ftgrid`-
  specific code, and the whole grid section is replaced with a tab view.
  Tabs are maintained using a separate `QVector` so we can easily access the
  "current tab". New singal `reloadCurrentTabFont`, `repaintCurrentTab` and
  the old `syncSettings` made use of this.

* src/ftinspect/panels/singular.hpp, src/ftinspect/panels/singular.cpp: New
  file, as described above.

* src/ftinspect/panels/abstracttab.hpp: New file, as described above.

* src/ftinspect/engine/engine.cpp, src/ftinspect/engine/engine.hpp: Save
  current font's glyph count so no need to pass it down from `MainGUI` to all
  tabs when a font reload happens.
  Add a dedicate `antiAliasingEnabled` property, which is set when anti-
  aliasing is set to "None".

* src/ftinspect/panels/settingpanel.cpp,
  src/ftinspect/panels/settingpanel.hpp: Move out all `ftgrid`-specific items.

* src/ftinspect/widgets/customwidgets.cpp,
  src/ftinspect/widgets/customwidgets.hpp: Renamed `QSpinBoxx` to
  `ZoomSpinBox` because "Zoom" box is its only usage and such name is much
  more descriptive.

* src/ftinspect/CMakeLists.txt, src/ftinspect/meson.build: Updated.