[graph] Rework cache management.
With more complex rendering in `ftgrid`, the old cache management
resulted in excessive clashes and slow `gblender_lookup`.  It is
better to replace the clashing entries than to preserve them and
search for an available spot in the table.  This saves about 10%
of execution time in `ftgrid` without much effect on the others.

* graph/gblender.c (gblender_lookup): Replace old clashing entries
and hash function.
(gblender_lookup_channel): Ditto.
* graph/gblender.h (GBlenderRec): Remove `stat_clears`.