[ftinspect] Refactor `Engine`.
This commit decouples `Engine` so it no longer depends on `MainGUI`.

Because of that, a lot of getters and setters are added to `Engine`, and
`MainGUI` call setters in its `applySettings` function, which makes applying
of the settings lazy (only happen when the glyph repaints).

In the later commits, options will be divided into two types: mostly lazy
ones, and a few eager ones which need to be immediately applied, because
they need the cache to be purged.

Some code in `MainGUI` is moved into `Engine`, as well.

Functions in the `Engine` are re-ordered. Constructor is greatly reduced to
`queryEngine`.

* src/ftinspect/engine/engine.cpp, src/ftinspect/engine/engine.hpp:
  As described.

* src/ftinspect/maingui.cpp, src/ftinspect/maingui.hpp:
  As described.

* src/ftinspect/ftinspect.cpp: Remove dependency of `Engine` against
  `MainGUI`.