[ftinspect] Refactor `Engine` and fix the singular tab.
This comment brings the new glyph loading and bitmap rendering code.
The bitmap rendering part is in `RenderingEngine` which can be obtained
from the `Engine`. The `Engine` now supports loading the glyph by loading
with image cache and without image cache.
The `loadOutline` function is removed.

Accompanying this change, some new features are introduced as well. Most
notable the improvements in rendering: back/foreground color and gamma,
support bitmap glyphs (not enabled in the UI). Support for LCD and color
layer font is absent yet.

The singular tab is now enabled by uncommented functional code that was
commented in the last commit.

In `Engine`, a `ftFallbackFace_` is introduced for all non-rendering work,
and the old `ftSize_` is only used for rendering. This helps dealing with
non-scalable fonts since we can safely retrieve info with the fallback face.

* src/ftinspect/engine/engine.cpp, src/ftinspect/engine/engine.hpp:
  Integrate the `RenderingEngine`.
  Change `loadFont` to update the `ftFallbackFace_` as well.
  Change all info-retrieving functions to use `ftFallbackFace_`.
  Add `loadGlyph`, `loadGlyphIntoSlotWithoutCache` and
  `loadGlyphWithoutUpdate`, and remove `loadOutline`.
  Extract code about "get a face object from the cache and do sth."
  into a new function `withFace`.
  Add property `antiAliasingEnabled` and `renderMode` for the rendering
  engine.
  Some minor changes are included.

* src/ftinspect/engine/rendering.cpp, src/ftinspect/engine/rendering.hpp:
  New files, as described.

* src/ftinspect/panels/settingpanel.cpp,
  src/ftinspect/glyphcomponents/glyphbitmap.cpp,
  src/ftinspect/panels/singular.cpp: Uncomment functional code.

* src/ftinspect/CMakeLists.txt, src/ftinspect/meson.build: Updated.