* src/ftinspect/panels/settingpanelmmgx.cpp: Fix layout.
Load a MM/GX font and switch to MM/GX tab in the left panel, then switch to
the comparator tab, the setting panels below will be abnormally tall due to
the updated MM/GX axes info loaded the panel with MM/GX axis widgets. This
commit fixes the issue.