[ftinspect] Fix uninitialized memory issue and reference misuse.
Fixes issue #23.

* src/ftinspect/engine/rendering.cpp (RenderingEngine::convertBitmapTo8Bpp):
Properly initialize `FT_Bitmap` since it is a value type.
(RenderingEngine::convertBitmapToQImage): Fix the accidental overwriting of
the source bitmap when converting the bitmap format (only happens for
`FT_PIXEL_MODE_GRAY2` and `FT_PIXEL_MODE_GRAY4`) due to a misuse of a C++
reference.

* src/ftinspect/glyphcomponents/glyphcontinuous.cpp
(GlyphContinuous::drawCacheGlyph): Fix a crash when the `QImage` was failed
to produce.